requirements.txt
opencv-python
flask
roboflow
numpy
qrcode
pillow
pymongo